fixes and tweaks (dirty)
fix time stop at open systemmap
fix too tall windows
add show indicators nav and combat target
add center body first click, select on second
can center on every body

+++ src/lua/LuaGame.cpp
@@ -617,7 +617,7 @@ static int l_game_get_hyperspace_travelled_percentage(lua_State *l)
 
 static int l_game_get_parts_from_date_time(lua_State *l)
 {
-	float time = LuaPull<float>(l, 1);
+	double time = LuaPull<double>(l, 1);
 	Time::DateTime t(time);
 	int year, month, day, hour, minute, second;
 	t.GetDateParts(&year, &month, &day);
